README fix for PHP 7.3 installations
The current installation guide will fail for PHP 7.3, so I added this fix: It tells the user to do a composer require to version `4.6`, which will install 4.6.0. This version has an [incompatibility issue](https://github.com/ARCANEDEV/LogViewer/issues/234) with PHP 7.3 that has been fixed in 4.6.1, so it makes sense to require that by default.

Maybe there is a better solution to require the *latest stable* of 4.6 rather than setting the .1 manually. Feel free to edit this commit :)